chore: remove in flight work for the next PR
