+++ src/notebooks/pipes/Visualization/types/Graph/index.tsx
@@ -1,110 +1,11 @@
-import React from 'react'
-import {DEFAULT_LINE_COLORS} from 'src/shared/constants/graphColorPalettes'
-import {
-  DEFAULT_FILLVALUES,
-  AGG_WINDOW_AUTO,
-} from 'src/timeMachine/constants/queryBuilder'
-
-import {Axis, Base, Color, XYViewProperties} from 'src/types'
-
-const icon = (
-  <div className="vis-graphic" data-testid="vis-graphic--xy">
-    <svg
-      width="100%"
-      height="100%"
-      version="1.1"
-      id="Line"
-      x="0px"
-      y="0px"
-      viewBox="0 0 150 150"
-      preserveAspectRatio="none meet"
-    >
-      <polygon
-        className="vis-graphic--fill vis-graphic--fill-a"
-        points="148,40 111.5,47.2 75,25 38.5,90.8 2,111.8 2,125 148,125 	"
-      />
-      <polyline
-        className="vis-graphic--line vis-graphic--line-a"
-        points="2,111.8 38.5,90.8 75,25 111.5,47.2 148,40 	"
-      />
-      <polygon
-        className="vis-graphic--fill vis-graphic--fill-b"
-        points="148,88.2 111.5,95.5 75,61.7 38.5,49.3 2,90.8 2,125 148,125 	"
-      />
-      <polyline
-        className="vis-graphic--line vis-graphic--line-b"
-        points="2,90.8 38.5,49.3 75,61.7 111.5,95.5 148,88.2 	"
-      />
-      <polygon
-        className="vis-graphic--fill vis-graphic--fill-c"
-        points="148,96 111.5,106.3 75,85.7 38.5,116.5 2,115 2,125 148,125 	"
-      />
-      <polyline
-        className="vis-graphic--line vis-graphic--line-c"
-        points="2,115 38.5,116.5 75,85.7 111.5,106.3 148,96 	"
-      />
-    </svg>
-  </div>
-)
+import icon from './icon'
+import properties from './properties'
 
 export default register => {
   register({
     type: 'xy',
     name: 'Graph',
     graphic: icon,
-    initial: {
-      type: 'xy',
-      shape: 'chronograf-v2',
-      geom: 'line',
-      xColumn: null,
-      yColumn: null,
-      position: 'overlaid',
-
-      queries: [
-        {
-          name: '',
-          text: '',
-          editMode: 'builder',
-          builderConfig: {
-            buckets: [],
-            tags: [
-              {
-                key: '_measurement',
-                values: [],
-                aggregateFunctionType: 'filter',
-              },
-            ],
-            functions: [{name: 'mean'}],
-            aggregateWindow: {
-              period: AGG_WINDOW_AUTO,
-              fillValues: DEFAULT_FILLVALUES,
-            },
-          },
-        },
-      ],
-
-      colors: DEFAULT_LINE_COLORS as Color[],
-      legend: {},
-      note: '',
-      showNoteWhenEmpty: false,
-      axes: {
-        x: {
-          bounds: ['', ''],
-          label: '',
-          prefix: '',
-          suffix: '',
-          base: '10',
-          scale: 'linear',
-        } as Axis,
-        y: {
-          bounds: ['', ''],
-          label: '',
-          prefix: '',
-          suffix: '',
-          base: '10' as Base,
-          scale: 'linear',
-        } as Axis,
-      },
-    } as XYViewProperties,
+    initial: properties,
   })
 }
