+++ src/notebooks/pipes/Visualization/view.tsx
@@ -106,10 +106,6 @@ export {_transform}
 const Visualization: FC<PipeProp> = ({Context}) => {
   const {timeZone} = useContext(AppSettingContext)
   const {data, update, loading, results} = useContext(PipeContext)
-  const [optionsVisibility, setOptionsVisibility] = useState(false)
-  const toggleOptions = useCallback(() => {
-    setOptionsVisibility(!optionsVisibility)
-  }, [optionsVisibility, setOptionsVisibility])
 
   const updateType = (type: ViewType) => {
     event('Notebook Visualization Type Changed', {
@@ -126,15 +122,6 @@ const Visualization: FC<PipeProp> = ({Context}) => {
       <ViewTypeDropdown
         viewType={data.properties.type}
         onUpdateType={updateType as any}
-      />
-      <SquareButton
-        icon={IconFont.CogThick}
-        onClick={toggleOptions}
-        color={
-          optionsVisibility ? ComponentColor.Primary : ComponentColor.Default
-        }
-        titleText="Configure Visualization"
-        className="flows-config-visualization-button"
       />
       <ExportVisualizationButton disabled={!results.source}>
         {onHidePopover => (
